Restructure cache controller into a package
Split the monolithic cache.py into a proper package matching the
pattern used by players, tasks, and streams controllers:

- controller.py: main CacheController class
- constants.py: shared constants, type aliases, context variables
- helpers.py: use_cache decorator
- __init__.py: re-exports for backwards compatibility
- README.md: documentation for contributors and LLMs
